requirements.txt
pyTelegramBotAPI
requests
gTTS
openai
groq-client
google-generativeai
jinja2
